Favoris + JSP + Taskqueue

+++ projetweb/src/main/java/com/projetweb/helper/ConstantesHelper.java
@@ -14,7 +14,7 @@ public class ConstantesHelper {
 	//Nombre de S dans une heure pour les conversion
 	public static final long S_IN_MINUTE = 60;
 	//Nombre de Mètres dans un km pour les conversion
-	public static final long M_IN_KM = 1000;
+	public static final int M_IN_KM = 1000;
 	//Nombre de minutes dans une heure
 	public static final int MINUTE_IN_HOUR = 60;
